fix(website): add delayed api fetching
There is a problem when you try to use a CORS public proxy, and that is the rate limit. For that reason a limit is imposed on API calls. The CORS proxy is only applied when the system recognizes a CORS block was applied to the request. A preferred solution would be a server dedicated to communicating with APIs, but the purpose of this application is to be able to work completely on the client side.